moved visual representation logic to game class as it is responsible for repersenting rthe appropriate clues as per game play
